linux/requirements.txt
altgraph==0.17.4
customtkinter==5.2.2
darkdetect==0.8.0
ewmh==0.1.6
MouseInfo==0.1.3
packaging==24.1
pillow==10.3.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
pyinstaller==6.8.0
pyinstaller-hooks-contrib==2024.7
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==0.1.30
python-xlib==0.33
python3-xlib==0.15
pytweening==1.2.0
setuptools==70.1.1
six==1.16.0
tk==0.1.0
